+++ packages/dataviews/src/field-types/date.tsx
@@ -6,14 +6,8 @@ import { dateI18n, getDate, getSettings } from '@wordpress/date';
 /**
  * Internal dependencies
  */
-import type {
-	DataViewRenderFieldProps,
-	Field,
-	FormatDate,
-	SortDirection,
-} from '../types';
+import type { FormatDate, NormalizedField, SortDirection } from '../types';
 import type { FieldType } from '../types/private';
-import RenderFromElements from './utils/render-from-elements';
 import isValidElements from './utils/is-valid-elements';
 import {
 	OPERATOR_ON,
@@ -25,49 +19,35 @@ import {
 	OPERATOR_IN_THE_PAST,
 	OPERATOR_OVER,
 	OPERATOR_BETWEEN,
-	DAYS_OF_WEEK,
 } from '../constants';
 import isValidRequired from './utils/is-valid-required';
+import render from './utils/render-default';
 
-function getFormat< Item >( field: Field< Item > ): Required< FormatDate > {
-	const fieldFormat = field.format as FormatDate | undefined;
-	return {
-		date:
-			fieldFormat?.date !== undefined &&
-			typeof fieldFormat.date === 'string'
-				? fieldFormat.date
-				: getSettings().formats.date,
-		weekStartsOn:
-			fieldFormat?.weekStartsOn !== undefined &&
-			DAYS_OF_WEEK.includes( fieldFormat?.weekStartsOn )
-				? fieldFormat.weekStartsOn
-				: getSettings().l10n.startOfWeek,
-	};
-}
+const format = {
+	date: getSettings().formats.date,
+	weekStartsOn: getSettings().l10n.startOfWeek,
+};
 
-function render( { item, field }: DataViewRenderFieldProps< any > ) {
-	if ( field.hasElements ) {
-		return <RenderFromElements item={ item } field={ field } />;
-	}
-
+function getValueFormatted< Item >( {
+	item,
+	field,
+}: {
+	item: Item;
+	field: NormalizedField< Item >;
+} ): string {
 	const value = field.getValue( { item } );
 	if ( [ '', undefined, null ].includes( value ) ) {
 		return '';
 	}
 
-	// If the field type is date, we've already normalized the format,
-	// and so it's safe to tell TypeScript to trust us ("as Required<Format>").
-	//
-	// There're no runtime paths where this render function is called with a non-date field,
-	// but TypeScript is unable to infer this, hence the type assertion.
-	let format: Required< FormatDate >;
+	let formatDate: Required< FormatDate >;
 	if ( field.type !== 'date' ) {
-		format = getFormat( {} as Field< any > );
+		formatDate = format;
 	} else {
-		format = field.format as Required< FormatDate >;
+		formatDate = field.format as Required< FormatDate >;
 	}
 
-	return dateI18n( format.date, getDate( value ) );
+	return dateI18n( formatDate.date, getDate( value ) );
 }
 
 const sort = ( a: any, b: any, direction: SortDirection ) => {
@@ -106,7 +86,8 @@ export default {
 		OPERATOR_OVER,
 		OPERATOR_BETWEEN,
 	],
-	getFormat,
+	format,
+	getValueFormatted,
 	validate: {
 		required: isValidRequired,
 		elements: isValidElements,