+++ packages/dataviews/src/field-types/datetime.tsx
@@ -6,14 +6,8 @@ import { dateI18n, getDate, getSettings } from '@wordpress/date';
 /**
  * Internal dependencies
  */
-import type {
-	DataViewRenderFieldProps,
-	Field,
-	FormatDatetime,
-	SortDirection,
-} from '../types';
+import type { FormatDatetime, NormalizedField, SortDirection } from '../types';
 import type { FieldType } from '../types/private';
-import RenderFromElements from './utils/render-from-elements';
 import isValidElements from './utils/is-valid-elements';
 import {
 	OPERATOR_ON,
@@ -24,33 +18,35 @@ import {
 	OPERATOR_AFTER_INC,
 	OPERATOR_IN_THE_PAST,
 	OPERATOR_OVER,
-	DAYS_OF_WEEK,
 } from '../constants';
 import isValidRequired from './utils/is-valid-required';
+import render from './utils/render-default';
 
-function render( { item, field }: DataViewRenderFieldProps< any > ) {
-	if ( field.elements ) {
-		return <RenderFromElements item={ item } field={ field } />;
-	}
+const format = {
+	datetime: getSettings().formats.datetime,
+	weekStartsOn: getSettings().l10n.startOfWeek,
+};
 
+function getValueFormatted< Item >( {
+	item,
+	field,
+}: {
+	item: Item;
+	field: NormalizedField< Item >;
+} ): string {
 	const value = field.getValue( { item } );
 	if ( [ '', undefined, null ].includes( value ) ) {
-		return null;
+		return '';
 	}
 
-	// If the field type is datetime, we've already normalized the format,
-	// and so it's safe to tell TypeScript to trust us ("as Required<Format>").
-	//
-	// There're no runtime paths where this render function is called with a non-datetime field,
-	// but TypeScript is unable to infer this, hence the type assertion.
-	let format: Required< FormatDatetime >;
+	let formatDatetime: Required< FormatDatetime >;
 	if ( field.type !== 'datetime' ) {
-		format = getFormat( {} as Field< any > );
+		formatDatetime = format;
 	} else {
-		format = field.format as Required< FormatDatetime >;
+		formatDatetime = field.format as Required< FormatDatetime >;
 	}
 
-	return dateI18n( format.datetime, getDate( value ) );
+	return dateI18n( formatDatetime.datetime, getDate( value ) );
 }
 
 const sort = ( a: any, b: any, direction: SortDirection ) => {
@@ -59,22 +55,6 @@ const sort = ( a: any, b: any, direction: SortDirection ) => {
 
 	return direction === 'asc' ? timeA - timeB : timeB - timeA;
 };
-
-function getFormat< Item >( field: Field< Item > ): Required< FormatDatetime > {
-	const fieldFormat = field.format as FormatDatetime | undefined;
-	return {
-		datetime:
-			fieldFormat?.datetime !== undefined &&
-			typeof fieldFormat.datetime === 'string'
-				? fieldFormat.datetime
-				: getSettings().formats.datetime,
-		weekStartsOn:
-			fieldFormat?.weekStartsOn !== undefined &&
-			DAYS_OF_WEEK.includes( fieldFormat?.weekStartsOn )
-				? fieldFormat.weekStartsOn
-				: getSettings().l10n.startOfWeek,
-	};
-}
 
 export default {
 	type: 'datetime',
@@ -103,7 +83,8 @@ export default {
 		OPERATOR_IN_THE_PAST,
 		OPERATOR_OVER,
 	],
-	getFormat,
+	format,
+	getValueFormatted,
 	validate: {
 		required: isValidRequired,
 		elements: isValidElements,